change: edit/delete via interaction endpoints
When the bot is editing/deleting it's message from an interaction, it now uses the interaction response endpoints to edit/delete messages, instead of using the normal messages endpoint. This also gives the added benefit of continuing to be able to use the components if the bot loses access to the channel.